refactor(providers): lift the key ring out of the Tavily provider
Multi-key rotation lived inside the Tavily provider, so the AI upstreams could
not use it without either duplicating the ring or reaching into a sibling
provider. Moving it to its own module lets the delimiter work and the AI-side
rotation proceed independently instead of queueing behind each other.

Nothing about the behaviour changes. tavily.rs gains exactly one line — the
import — and loses everything else; the parse rules, the randomized start, the
shared cursor and the key-scoped status set are byte-identical, and the
existing rotation integration tests pass with their assertions untouched.

The KeyRing unit tests move next to the ring so the same assertions do not run
twice. The one test that stays behind is about how the provider wires the ring
up, which is the provider's own behaviour rather than the ring's.

Tavily's 432/433 stay in the key-scoped set. They are inert for upstreams that
never return them, so one predicate serves every provider; only the comment
changed to say so.

The module is pub(crate); the public API surface is unchanged.
