+++ src/providers/mod.rs
@@ -2,6 +2,8 @@ pub mod exa;
 pub mod firecrawl;
 pub mod grok;
 pub mod http;
+/// Crate-internal: shared by the providers, not part of the public surface.
+pub(crate) mod keyring;
 pub mod openai_compatible;
 pub mod tavily;
 pub mod tinyfish;
